refactor(lox-analysis): split imaging module into submodules
