+++ crates/lox-analysis/src/imaging.rs
@@ -2,11 +2,19 @@
 //
 // SPDX-License-Identifier: MPL-2.0
 
-use std::collections::HashMap;
-use std::fmt;
+//! AOI access analysis: sensor payloads, AOI geometry, and per-(spacecraft, AOI) access windows.
 
-use geo::prelude::Contains;
-use geo::{Closest, HaversineClosestPoint};
+pub mod aoi;
+/// Imaging analysis result types.
+pub mod results;
+
+pub use aoi::{Aoi, AoiId};
+pub use results::ImagingResults;
+
+#[cfg(feature = "geojson")]
+pub use aoi::AoiError;
+
+use std::collections::HashMap;
 
 use lox_bodies::{DynOrigin, Origin, TryMeanRadius, TrySpheroid};
 use lox_core::coords::LonLatAlt;
@@ -20,177 +28,12 @@ use lox_orbits::events::{
 use lox_orbits::orbits::{Ensemble, Trajectory};
 use lox_time::Time;
 use lox_time::deltas::TimeDelta;
-use lox_time::intervals::TimeInterval;
 use lox_time::time_scales::Tai;
 use rayon::prelude::*;
 use thiserror::Error;
 
 use crate::assets::{AssetId, Scenario, Spacecraft};
 use crate::visibility::EvalError;
-
-/// Densify a polygon by inserting intermediate vertices via linear lon/lat
-/// interpolation so that no sub-segment exceeds `max_deg` degrees (Euclidean
-/// distance in lon/lat coordinate space).
-///
-/// `geo::HaversineClosestPoint` treats each polygon edge as a great-circle arc.
-/// For long edges (e.g. 100° of longitude at mid-latitudes) that arc can curve
-/// many degrees away from the intended constant-latitude boundary, producing
-/// near-zero cross-track distances for satellites that are in fact far outside
-/// the AOI. By keeping sub-segments ≤ 0.5° long the great-circle arc and the
-/// straight lon/lat line are indistinguishable for practical sensor ranges.
-fn densify_polygon_linear(polygon: geo::Polygon<f64>, max_deg: f64) -> geo::Polygon<f64> {
-    let exterior = densify_ring_linear(polygon.exterior(), max_deg);
-    let interiors: Vec<_> = polygon
-        .interiors()
-        .iter()
-        .map(|ring| densify_ring_linear(ring, max_deg))
-        .collect();
-    geo::Polygon::new(exterior, interiors)
-}
-
-fn densify_ring_linear(ring: &geo::LineString<f64>, max_deg: f64) -> geo::LineString<f64> {
-    let coords: Vec<geo::Coord<f64>> = ring.coords().copied().collect();
-    let mut out: Vec<geo::Coord<f64>> = Vec::with_capacity(coords.len() * 4);
-    for pair in coords.windows(2) {
-        let (x0, y0) = (pair[0].x, pair[0].y);
-        let (x1, y1) = (pair[1].x, pair[1].y);
-        let dx = x1 - x0;
-        let dy = y1 - y0;
-        let len = (dx * dx + dy * dy).sqrt();
-        let n = ((len / max_deg).ceil() as usize).max(1);
-        out.push(geo::Coord { x: x0, y: y0 });
-        for i in 1..n {
-            let t = i as f64 / n as f64;
-            out.push(geo::Coord {
-                x: x0 + t * dx,
-                y: y0 + t * dy,
-            });
-        }
-    }
-    if let Some(&last) = coords.last() {
-        out.push(last);
-    }
-    geo::LineString::new(out)
-}
-
-/// Haversine great-circle distance between two lon/lat points (degrees) on a
-/// sphere of the given radius (meters).
-fn haversine_distance(a: geo::Point<f64>, b: geo::Point<f64>, radius_m: f64) -> f64 {
-    let (lon1, lat1) = (a.x().to_radians(), a.y().to_radians());
-    let (lon2, lat2) = (b.x().to_radians(), b.y().to_radians());
-    let dlat = lat2 - lat1;
-    let dlon = lon2 - lon1;
-    let h = (dlat / 2.0).sin().powi(2) + lat1.cos() * lat2.cos() * (dlon / 2.0).sin().powi(2);
-    2.0 * radius_m * h.sqrt().asin()
-}
-
-// ---------------------------------------------------------------------------
-// AoiId
-// ---------------------------------------------------------------------------
-
-/// Unique identifier for an area of interest.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct AoiId(String);
-
-impl AoiId {
-    /// Creates a new AOI identifier.
-    pub fn new(id: impl Into<String>) -> Self {
-        Self(id.into())
-    }
-
-    /// Returns the identifier as a string slice.
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl fmt::Display for AoiId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Aoi — Area of Interest
-// ---------------------------------------------------------------------------
-
-/// An area of interest defined as a geographic polygon.
-///
-/// Coordinates follow GeoJSON convention: longitude/latitude in degrees.
-#[derive(Debug, Clone)]
-pub struct Aoi {
-    polygon: geo::Polygon<f64>,
-}
-
-/// Errors from AOI construction.
-#[cfg(feature = "geojson")]
-#[derive(Debug, Error)]
-pub enum AoiError {
-    /// Invalid GeoJSON string.
-    #[error("invalid GeoJSON: {0}")]
-    InvalidGeoJson(String),
-    /// No polygon found in the GeoJSON.
-    #[error("no polygon found in GeoJSON")]
-    NoPolygon,
-}
-
-impl Aoi {
-    /// Creates a new AOI from a `geo::Polygon` (lon/lat in degrees).
-    pub fn new(polygon: geo::Polygon<f64>) -> Self {
-        let polygon = densify_polygon_linear(polygon, 0.5);
-        Self { polygon }
-    }
-
-    /// Parses an AOI from a GeoJSON string.
-    ///
-    /// Expects a GeoJSON Polygon geometry (or a Feature containing one).
-    #[cfg(feature = "geojson")]
-    pub fn from_geojson(geojson: &str) -> Result<Self, AoiError> {
-        use geojson::GeoJson;
-        use std::convert::TryInto;
-
-        let gj: GeoJson = geojson
-            .parse()
-            .map_err(|e: geojson::Error| AoiError::InvalidGeoJson(e.to_string()))?;
-
-        let geometry = match gj {
-            GeoJson::Geometry(g) => g,
-            GeoJson::Feature(f) => f.geometry.ok_or(AoiError::NoPolygon)?,
-            GeoJson::FeatureCollection(fc) => fc
-                .features
-                .into_iter()
-                .find_map(|f| f.geometry)
-                .ok_or(AoiError::NoPolygon)?,
-        };
-
-        let polygon: geo::Polygon<f64> =
-            geometry.value.try_into().map_err(|_| AoiError::NoPolygon)?;
-
-        Ok(Self::new(polygon))
-    }
-
-    /// Returns the great-circle distance in meters from a point to the AOI polygon.
-    ///
-    /// Uses the haversine formula with the given `mean_radius_m` so that the
-    /// computation is valid for any spherical body, not just Earth.
-    /// Returns 0.0 if the point is inside the polygon.
-    pub fn distance_to(&self, point: &geo::Point<f64>, mean_radius_m: f64) -> f64 {
-        if self.polygon.contains(point) {
-            0.0
-        } else {
-            match self.polygon.haversine_closest_point(point) {
-                Closest::Intersection(_) => 0.0,
-                Closest::SinglePoint(closest) => haversine_distance(*point, closest, mean_radius_m),
-                Closest::Indeterminate => unreachable!("degenerate polygon geometry"),
-            }
-        }
-    }
-
-    /// Returns a reference to the underlying polygon.
-    pub fn polygon(&self) -> &geo::Polygon<f64> {
-        &self.polygon
-    }
-}
 
 // ---------------------------------------------------------------------------
 // ImagingPayload — Imaging sensor payload
@@ -321,42 +164,6 @@ pub enum ImagingError {
 }
 
 // ---------------------------------------------------------------------------
-// ImagingResults
-// ---------------------------------------------------------------------------
-
-type ImagingIntervalMap = HashMap<(AssetId, AoiId), Vec<TimeInterval<Tai>>>;
-
-/// Results of an imaging analysis.
-pub struct ImagingResults {
-    intervals: ImagingIntervalMap,
-}
-
-impl ImagingResults {
-    /// Returns imaging intervals for a specific (spacecraft, AOI) pair.
-    pub fn intervals(&self, sc_id: &AssetId, aoi_id: &AoiId) -> &[TimeInterval<Tai>] {
-        self.intervals
-            .get(&(sc_id.clone(), aoi_id.clone()))
-            .map(|v| v.as_slice())
-            .unwrap_or(&[])
-    }
-
-    /// Returns an iterator over all (spacecraft, AOI) pairs and their intervals.
-    pub fn all_intervals(&self) -> &ImagingIntervalMap {
-        &self.intervals
-    }
-
-    /// Returns `true` if no imaging intervals were found.
-    pub fn is_empty(&self) -> bool {
-        self.intervals.is_empty()
-    }
-
-    /// Returns the number of (spacecraft, AOI) pairs.
-    pub fn num_pairs(&self) -> usize {
-        self.intervals.len()
-    }
-}
-
-// ---------------------------------------------------------------------------
 // ImagingAnalysis — orchestration
 // ---------------------------------------------------------------------------
 
@@ -454,192 +261,15 @@ where
         };
 
         let intervals: HashMap<_, _> = results?.into_iter().collect();
-        Ok(ImagingResults { intervals })
+        Ok(ImagingResults::new(intervals))
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use geo::{LineString, Polygon, point};
-    use lox_bodies::{Earth, MeanRadius};
-
-    fn earth_mean_radius_m() -> f64 {
-        Earth.mean_radius().to_meters()
-    }
-
-    #[test]
-    fn test_aoi_id() {
-        let id = AoiId::new("rome");
-        assert_eq!(id.as_str(), "rome");
-        assert_eq!(format!("{id}"), "rome");
-    }
-
-    #[test]
-    fn test_aoi_distance_inside() {
-        let polygon = Polygon::new(
-            LineString::from(vec![
-                (10.0, 45.0),
-                (11.0, 45.0),
-                (11.0, 46.0),
-                (10.0, 46.0),
-                (10.0, 45.0),
-            ]),
-            vec![],
-        );
-        let aoi = Aoi::new(polygon);
-        let inside = point!(x: 10.5, y: 45.5);
-        assert_eq!(aoi.distance_to(&inside, earth_mean_radius_m()), 0.0);
-    }
-
-    // Regression test for issue #401: large AOI at high latitudes produced false
-    // imaging intervals because geo::HaversineClosestPoint treats polygon edges as
-    // great-circle arcs. The great circle connecting (-45°,50°) to (55°,50°)
-    // curves up to ~61.7°N at its midpoint (lon≈5°). A satellite at (5°,62°N)
-    // therefore appeared to be only ~38 km from the polygon boundary, even though
-    // the true distance to the lat=50° edge is ~1334 km. With a sensor range of
-    // ~290 km this produced a false imaging detection.
-    #[test]
-    fn test_aoi_distance_large_polygon_near_great_circle_arc_peak() {
-        let polygon = Polygon::new(
-            LineString::from(vec![
-                (-45.0, 30.0),
-                (55.0, 30.0),
-                (55.0, 50.0),
-                (-45.0, 50.0),
-                (-45.0, 30.0),
-            ]),
-            vec![],
-        );
-        let aoi = Aoi::new(polygon);
-        // 12° north of the top edge – the great-circle arc of that edge peaks at
-        // ~61.7°N, so (5°,62°) sits almost on the arc, yielding a near-zero
-        // computed distance before the fix.
-        let near_arc_peak = point!(x: 5.0, y: 62.0);
-        let d = aoi.distance_to(&near_arc_peak, earth_mean_radius_m());
-        // True distance to the 50°N boundary is ~1334 km.
-        // Before the fix this returned ~38 km (false positive).
-        assert!(
-            d > 1_000_000.0,
-            "expected distance > 1000 km for point 12° north of AOI boundary, got {:.0} m",
-            d
-        );
-    }
-
-    #[test]
-    fn test_aoi_distance_outside() {
-        let polygon = Polygon::new(
-            LineString::from(vec![
-                (10.0, 45.0),
-                (11.0, 45.0),
-                (11.0, 46.0),
-                (10.0, 46.0),
-                (10.0, 45.0),
-            ]),
-            vec![],
-        );
-        let aoi = Aoi::new(polygon);
-        let outside = point!(x: 12.0, y: 45.5);
-        let d = aoi.distance_to(&outside, earth_mean_radius_m());
-        assert!(d > 0.0);
-        // ~78 km from edge
-        assert!(d > 70_000.0 && d < 90_000.0);
-    }
-
-    #[test]
-    fn test_haversine_matches_geodesic_for_earth() {
-        use geo::{Distance as GeoDistance, Geodesic};
-
-        // Two points ~78 km apart (edge of AOI to a point outside)
-        let a = point!(x: 11.0, y: 45.5); // on polygon edge
-        let b = point!(x: 12.0, y: 45.5); // 1° east
-
-        let geodesic = Geodesic::distance(a, b);
-        let haversine = haversine_distance(a, b, earth_mean_radius_m());
-
-        // Haversine on a sphere vs Vincenty on WGS-84 ellipsoid: expect < 0.5% error
-        let rel_err = (haversine - geodesic).abs() / geodesic;
-        assert!(
-            rel_err < 0.005,
-            "haversine ({haversine:.1}) vs geodesic ({geodesic:.1}): {:.2}% error",
-            rel_err * 100.0,
-        );
-    }
-
-    #[test]
-    fn test_distance_scales_with_body_radius() {
-        let polygon = Polygon::new(
-            LineString::from(vec![
-                (10.0, 45.0),
-                (11.0, 45.0),
-                (11.0, 46.0),
-                (10.0, 46.0),
-                (10.0, 45.0),
-            ]),
-            vec![],
-        );
-        let aoi = Aoi::new(polygon);
-        let outside = point!(x: 12.0, y: 45.5);
-
-        let earth_d = aoi.distance_to(&outside, earth_mean_radius_m());
-        // Mars mean radius ~3389.5 km
-        let mars_d = aoi.distance_to(&outside, 3_389_500.0);
-
-        // Same angular separation, smaller body → shorter distance
-        assert!(mars_d < earth_d);
-        let ratio = earth_d / mars_d;
-        let expected_ratio = earth_mean_radius_m() / 3_389_500.0;
-        assert!(
-            (ratio - expected_ratio).abs() < 0.01,
-            "distance should scale linearly with radius"
-        );
-    }
-
-    #[test]
-    fn test_imaging_params_nadir_only() {
-        let params = ImagingPayload::nadir_only(Distance::kilometers(20.0));
-        assert_eq!(params.half_swath_ground_range, 10_000.0);
-        assert_eq!(params.max_off_nadir, 0.0);
-        // Nadir-only: range is just half swath
-        let range = params.max_accessible_ground_range(500_000.0, 6_371_000.0);
-        assert!((range - 10_000.0).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_imaging_params_off_nadir() {
-        let params = ImagingPayload::off_nadir(Distance::kilometers(20.0), Angle::degrees(30.0));
-        let range = params.max_accessible_ground_range(500_000.0, 6_371_000.0);
-        // Should be > half swath (10 km) due to off-nadir contribution
-        assert!(range > 10_000.0);
-        // For 30° off-nadir at 500 km altitude, expect ~300-400 km total
-        assert!(range > 200_000.0);
-    }
-
-    #[cfg(feature = "geojson")]
-    #[test]
-    fn test_aoi_from_geojson() {
-        let geojson =
-            r#"{"type":"Polygon","coordinates":[[[10,45],[11,45],[11,46],[10,46],[10,45]]]}"#;
-        let aoi = Aoi::from_geojson(geojson).unwrap();
-        let inside = point!(x: 10.5, y: 45.5);
-        assert_eq!(aoi.distance_to(&inside, earth_mean_radius_m()), 0.0);
-    }
-
-    #[cfg(feature = "geojson")]
-    #[test]
-    fn test_aoi_from_geojson_feature() {
-        let geojson = r#"{"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[10,45],[11,45],[11,46],[10,46],[10,45]]]},"properties":{}}"#;
-        let aoi = Aoi::from_geojson(geojson).unwrap();
-        let inside = point!(x: 10.5, y: 45.5);
-        assert_eq!(aoi.distance_to(&inside, earth_mean_radius_m()), 0.0);
-    }
-
-    #[cfg(feature = "geojson")]
-    #[test]
-    fn test_aoi_from_geojson_invalid() {
-        let result = Aoi::from_geojson("not json");
-        assert!(result.is_err());
-    }
+    use geo::{LineString, Polygon};
+    use lox_time::intervals::TimeInterval;
 
     // -----------------------------------------------------------------------
     // Integration tests — full ImagingAnalysis pipeline with Sentinel-2 TLEs
@@ -739,6 +369,26 @@ mod tests {
             ]),
             vec![],
         ))
+    }
+
+    #[test]
+    fn test_imaging_params_nadir_only() {
+        let params = ImagingPayload::nadir_only(Distance::kilometers(20.0));
+        assert_eq!(params.half_swath_ground_range, 10_000.0);
+        assert_eq!(params.max_off_nadir, 0.0);
+        // Nadir-only: range is just half swath
+        let range = params.max_accessible_ground_range(500_000.0, 6_371_000.0);
+        assert!((range - 10_000.0).abs() < 1e-6);
+    }
+
+    #[test]
+    fn test_imaging_params_off_nadir() {
+        let params = ImagingPayload::off_nadir(Distance::kilometers(20.0), Angle::degrees(30.0));
+        let range = params.max_accessible_ground_range(500_000.0, 6_371_000.0);
+        // Should be > half swath (10 km) due to off-nadir contribution
+        assert!(range > 10_000.0);
+        // For 30° off-nadir at 500 km altitude, expect ~300-400 km total
+        assert!(range > 200_000.0);
     }
 
     #[test]
